+++ src/gui.js
@@ -91,7 +91,7 @@ modules.gui = '2024-February-20';
 
 // Declarations
 
-var SnapVersion = '9.2.7-dev';
+var SnapVersion = '9.2.7';
 
 var IDE_Morph;
 var ProjectDialogMorph;