+++ HISTORY.md
@@ -2,8 +2,12 @@
 
 ## in development:
 
+* **New Features:**
+    * new "Lirin" costume series, thanks, Meghan and Brian!
+
 2024-02-20
 * new dev version for v9.2.7
+* added new Lirin costumes, thanks, Brian and Meghan!
 
 ## 9.2.6:
 * **Notable Changes:**